+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [1.1.5] - Unreleased
 - Add Color::contrast, grayscale and rgb_color.
+- Add doc comment that SvgFileSurface can panic on creation.
 
 ## [1.1.4] - 2021-07-23
 - Add app::set_frame_type_cb() and set_frame2().